requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
dnspython==2.4.2
idna==3.4
loguru==0.7.2
numpy==1.26.1
Pillow==10.1.0
pinecone-client==2.2.4
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
